refactor: simplify test page
